Simple Agent-Based Model
Creates a 20x20 grid of 0s and 1s, which represent values of some
variable held by agents in those cells. It then chooses two adjacent
cells, the first at random and the second at random from among the
first cell's neighbors, and applies a simple rule -- the first cell
takes on the value of the second. It iterates this cell selection and
rule application 1,000 times, displays the result, and tracks the
fraction of 1s in the matrix over time.