Replication of Axelrod's Dissemination of Culture model
Specifies a world of x by x dimensions. Each cell is occupied by an agent with y features and z traits per feature. In each step, randomly select a cell and have it interact with a randomly selected Von Neumann neighbor. Set the probability of interaction p equal to the percent of feature/trait pairs between the two agents that match. If the two cells do interact, have the selected cell take on the first feature/trait of their selected neighbor that doesn't already match.