Miller and Page's Standing Ovation Problem
Models a standing ovation. Sets an audience of size x by x. Each agent in the audience perceives the performance to have a certain quality with mean y and variance yvar. Each agent has a quality threshold for standing with mean z and variance zvar. Each agent has a social threshold for standing with mean s and variance svar.

When the performance ends, audience members stand if quality > quality threshold. In each round, audience members who are not already standing look around based on a vision rule (the default randomly selects between the five neighbors to their immediate left/right/front or the entire audience, but one or the other can be selected for all agents), and if the percentage of agents they see standing exceeds their social threshold they stand.